Add files via upload
Also, if anyone is wondering why it says "above" on the 1st step, that's because when I did this project's ".sb2", it originally was made in the 2.0 beta so it looked different from the final 2.0, but I decided to make it that. However I swapped it to below when I did that.